generator/requirements.txt
azure-ai-inference==1.0.0b9
python-dotenv==1.1.0